feat: allow clients to pick gamemodes
Allow clients to request a gamemode and bot count when registering and
only run games of that nature.

This does not quite work correctly, as if the first bot requests 'value'
and the second requests 'collection', it will be down to whoever
acquires the lock first. Additionally, if the first requests 4 bots and
the second requests 2, the game will begin immediately.