feat: allow multiple rooms
When users connect to the websocket, move them to the room they request
and only begin games when each room has enough bots. Once all websockets
have been informed of the winners, delete the room so it can be used
again.

This allows clients to specify a gametype and room size and be matched
up only when enough people choose the same settings.