Add option to run multiple auctions
Adds MultiRoom and SingleRoom classes which contains the bots and the
winers. The RoomKey now contains a roomtype either SingleRoom or MultiRoom
so the rooms are split. The number of runs is stored in the room and will
be the minimum number of requested rooms, > 1.
A new message is sent that has a dictionary of names and says how many
each bot won.